基本完成,但问题和c#一致

+++ UFunc.cpp
@@ -5,7 +5,7 @@
 #include "UFunc.h"
 
 namespace UnPdeC {
-	std::uint32_t UFunc::Get4Byte(const std::vector<std::uint8_t>& data, std::size_t offset) {
+	std::uint32_t UFunc::Get4Byte(const std::vector<std::uint8_t>& data, std::uint32_t offset) {
 		if (offset + 4 > data.size()) {
 			throw std::out_of_range("Offset out of range");
 		}
